@@ -1,32 +1,31 @@
-source 'https://rubygems.org'
+source "https://rubygems.org"
 
-ruby '2.2.3'
+ruby "2.3.0"
 
-gem 'rails', '4.2.4'
-gem 'pg', '0.18.1'
+gem "rails", "4.2.6"
+gem "pg", "0.18.1"
 
-gem 'responders', '~> 2.0.0'
+gem "responders"
 
-gem 'slim-rails', '~> 3.0.1'
+gem "slim-rails", "~> 3.0"
 
-gem 'redcarpet', '~> 3.3.2'
+gem "redcarpet"
 
-gem 'sass-rails', '~> 5.0.1'
-gem 'bootstrap-sass', '~> 3.3.3'
+gem "sass-rails", "~> 5.0"
+gem "bootstrap-sass", "~> 3.3"
 
 group :development do
-  gem 'spring'
+  gem "spring"
 end
 
 group :development, :test do
-  gem 'rspec-rails', '~> 3.1.0'
-  gem 'factory_girl_rails', '~> 4.5.0'
-  gem 'database_cleaner', '~> 1.4.0'
-  gem 'shoulda-matchers', '~> 2.7.0'
-  gem 'rubocop', '~> 0.32.1', require: false
-  gem 'dotenv-rails', '~> 2.0.1'
+  gem "rspec-rails"
+  gem "factory_girl_rails"
+  gem "shoulda-matchers"
+  gem "rubocop", require: false
+  gem "dotenv-rails"
 end
 
 group :production do
-  gem 'rails_12factor'
+  gem "rails_12factor"
 end
