@@ -1,3 +1,3 @@
 ActiveSupport.on_load(:action_controller) do
-  wrap_parameters :format => [:json] if respond_to?(:wrap_parameters)
+  wrap_parameters format: [:json] if respond_to?(:wrap_parameters)
 end
