@@ -2,5 +2,5 @@ class Post < ActiveRecord::Base
   has_many :authorships
   has_many :authors, through: :authorships
 
-  validates_presence_of :slug
+  validates :slug, presence: true
 end
