@@ -1,3 +1,3 @@
-require File.expand_path('../application', __FILE__)
+require File.expand_path("../application", __FILE__)
 
 Rails.application.initialize!
