@@ -1,1 +1,1 @@
-Rails.application.config.assets.version = '1.0'
+Rails.application.config.assets.version = "1.0"
