@@ -1,17 +1,17 @@
-require 'rails_helper'
+require "rails_helper"
 
-describe Github::File do
-  subject(:file) { described_class.new('foo.md') }
+RSpec.describe Github::File do
+  subject(:file) { described_class.new("foo.md") }
 
-  describe '#name' do
-    it { expect(file.name).to eq('foo.md') }
+  describe "#name" do
+    it { expect(file.name).to eq("foo.md") }
   end
 
-  describe '#slug' do
-    it { expect(file.slug).to eq('foo') }
+  describe "#slug" do
+    it { expect(file.slug).to eq("foo") }
   end
 
-  describe '#uri' do
-    it { expect(file.uri.to_s).to eq("#{ENV['POSTS_URL']}/master/foo.md") }
+  describe "#uri" do
+    it { expect(file.uri.to_s).to eq("#{ENV["POSTS_URL"]}/master/foo.md") }
   end
 end
