@@ -1,4 +1,4 @@
-require 'rails_helper'
+require "rails_helper"
 
 RSpec.describe Author do
   it { should have_many(:posts).through(:authorships) }
