@@ -1,4 +1,4 @@
-require 'rails_helper'
+require "rails_helper"
 
 RSpec.describe Post do
   it { should have_many(:authors).through(:authorships) }
