@@ -1,5 +1,5 @@
-require 'rails_helper'
+require "rails_helper"
 
-describe PostsController do
-  it { expect(get: 'posts/on-the-road').to route_to('posts#show', slug: 'on-the-road') }
+RSpec.describe PostsController do
+  it { expect(get: "posts/on-the-road").to route_to("posts#show", slug: "on-the-road") }
 end
