@@ -1,5 +1,5 @@
-require 'rails_helper'
+require "rails_helper"
 
-describe Webhooks::GithubController do
-  it { expect(post: 'webhooks/github/push_event').to route_to('webhooks/github#push_event') }
+RSpec.describe Webhooks::GithubController do
+  it { expect(post: "webhooks/github/push_event").to route_to("webhooks/github#push_event") }
 end
